# quick-pki-cert-api/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Quick-PKI Certificate API
  version: "1.0.0"
  description: |
    A RESTful certificate issuance service. A caller submits a base64-encoded
    PKCS#10 CSR and receives a signed certificate back in the same encoding.

    Routes under `/v1` are an OAuth 2.0 protected resource: every request must
    present a bearer access token, validated by RFC 7662 token introspection
    against an externally configured authorization server. Callers obtain their
    tokens using the client credentials grant. When the service is configured
    with a required scope, the token must additionally carry that scope.

    The issuer certificate, health, and documentation endpoints are
    unauthenticated.
servers:
  - url: __SERVER_URL__
    description: This deployment
...
paths:
  /v1/certificates:
    post:
      tags: [Certificates]
      summary: Issue a certificate from a CSR
      description: |
        Signs the supplied PKCS#10 certificate request with the issuing CA,
        persists the result, and returns it. The `Location` response header
        points at the persisted certificate resource.
      operationId: issueCertificate
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CertificateRequest'
      responses:
        '201':
          description: Certificate issued
          headers:
            Location:
              description: URL of the persisted certificate resource
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssuedCertificate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /v1/openssl-configs/brcac:
    post:
      tags: [OpenSSL configs]
      summary: Generate an openssl req config for an Open Finance Brasil BRCAC CSR
      description: |
        Returns an openssl `req` config that bakes in the BRCAC subject DN
        (in the spec-mandated RDN order) and the requested DNS SANs. The
        caller saves the config to disk and produces a compliant CSR locally
        with their own private key:

        ```
        openssl req -new -config brcac.cnf -key my-key.pem -out my-csr.pem
        ```

        The resulting CSR can then be submitted to `POST /v1/certificates`
        with `profile=BRCAC` for issuance.
      operationId: generateBrcacOpensslConfig
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BrcacOpensslConfigRequest'
      responses:
        '200':
          description: openssl config generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpensslConfigResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /v1/openssl-configs/brseal:
    post:
      tags: [OpenSSL configs]
      summary: Generate an openssl req config for an Open Finance Brasil BRSEAL CSR
      description: |
        BRSEAL counterpart to `/v1/openssl-configs/brcac`. Returns an openssl
        `req` config with the BRSEAL subject DN, three organizationalUnit
        RDNs, and the four required ICP-Brasil otherName SAN entries. The
        resulting CSR can be submitted to `POST /v1/certificates` with
        `profile=BRSEAL` for issuance.
      operationId: generateBrsealOpensslConfig
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BrsealOpensslConfigRequest'
      responses:
        '200':
          description: openssl config generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpensslConfigResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /v1/openssl-configs/qwac:
    post:
      tags: [OpenSSL configs]
      summary: Generate an openssl req config for an EU PSD2 QWAC CSR
      description: |
        Returns an openssl `req` config that bakes in the QWAC subject DN
        (in the ETSI EN 319 412-1 RDN order), the requested DNS SANs, the two
        mandatory ETSI QC statements (QcCompliance + QcType=web), and -- when
        `psd2Roles` is supplied -- the PSD2 qcStatement listing the PSP roles
        and supervising NCA. The resulting CSR can be submitted to
        `POST /v1/certificates` with `profile=QWAC` for issuance.
      operationId: generateQwacOpensslConfig
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QwacOpensslConfigRequest'
      responses:
        '200':
          description: openssl config generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpensslConfigResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /v1/openssl-configs/qseal:
    post:
      tags: [OpenSSL configs]
      summary: Generate an openssl req config for an EU PSD2 QSEAL CSR
      description: |
        QSEAL counterpart to `/v1/openssl-configs/qwac`. Returns an openssl
        `req` config with the QSEAL subject DN, no DNS SANs, no
        ExtendedKeyUsage, and the mandatory ETSI QC statements
        (QcCompliance + QcType=eseal). The resulting CSR can be submitted to
        `POST /v1/certificates` with `profile=QSEAL` for issuance.
      operationId: generateQsealOpensslConfig
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QsealOpensslConfigRequest'
      responses:
        '200':
          description: openssl config generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpensslConfigResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /v1/certificates/{id}:
    get:
      tags: [Certificates]
      summary: Fetch a previously issued certificate
      operationId: getCertificate
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The certificate's UUID, as returned by the issue call.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The certificate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssuedCertificate'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: OAuth 2.0 access token
      description: |
        An OAuth 2.0 bearer access token obtained via the client credentials
        grant from the authorization server this service is configured to
        trust.
  schemas:
...
    CertificateRequest:
      type: object
      required: [csr]
      properties:
        csr:
          type: string
          format: byte
          description: |
            A base64 encoding of a PKCS#10 certificate request. The decoded
            bytes may be either raw DER or a PEM-wrapped request.
          example: LS0tLS1CRUdJTiBDRVJUSUZJQ0FURSBSRVFVRVNULS0tLS0K
        profile:
          type: string
          enum: [DEFAULT, TLS_SERVER, TLS_CLIENT, BRCAC, BRSEAL, QWAC, QSEAL]
          default: DEFAULT
          description: |
            The certificate profile to issue under. The profile fixes the
            leaf's KeyUsage and ExtendedKeyUsage. `BRCAC` and `BRSEAL` are the
            Open Finance Brasil transport and signing certificate profiles;
            `QWAC` and `QSEAL` are the EU eIDAS / PSD2 Qualified Web
            Authentication and Qualified Electronic Seal profiles
            (ETSI EN 319 412-2/-3, ETSI TS 119 495 for the PSD2 attribute
            set). For any non-DEFAULT profile the CSR subject, subjectAltName
            extension, and -- for `QWAC` / `QSEAL` -- the requested
            qCStatements extension must carry the attributes the profile
            requires; non-compliant CSRs are rejected with `bad_csr`.
            Omitted or `DEFAULT` issues a general-purpose TLS leaf.
          example: BRCAC
    BrcacOpensslConfigRequest:
      type: object
      required:
        - commonName
        - businessCategory
        - serialNumber
        - organization
        - stateOrProvince
        - locality
        - organizationIdentifier
        - userId
        - dnsNames
      properties:
        commonName:
          type: string
          description: Subject commonName; typically the transport hostname.
          example: transport.example.com
        businessCategory:
          type: string
          description: EV businessCategory.
          enum:
            - Private Organization
            - Government Entity
            - Business Entity
            - Non-Commercial Entity
        serialNumber:
          type: string
          description: Subject serialNumber RDN; carries the company CNPJ.
          example: "12345678000199"
        organization:
          type: string
          example: Example Participant Ltda
        stateOrProvince:
          type: string
          example: SP
        locality:
          type: string
          example: Sao Paulo
        organizationIdentifier:
          type: string
          description: Participant code, prefixed with `OFBBR-`.
          example: OFBBR-12345678
        userId:
          type: string
          description: Software-statement identifier carried as the UID RDN.
        country:
          type: string
          default: BR
          description: Subject country; defaults to `BR`.
        jurisdictionCountry:
          type: string
          default: BR
          description: Jurisdiction-of-incorporation country; defaults to `BR`.
        dnsNames:
          type: array
          items:
            type: string
          description: DNS subjectAltName entries; at least one is required.
          example:
            - transport.example.com
    BrsealOpensslConfigRequest:
      type: object
      required:
        - commonName
        - userId
        - organizationUnits
        - responsiblePersonName
        - companyCnpj
        - responsiblePersonData
        - companyCei
      properties:
        commonName:
          type: string
          example: Seal Co
        userId:
          type: string
          description: UID RDN; participant code prefixed with `OFBBR-`.
          example: OFBBR-12345678
        country:
          type: string
          default: BR
        organization:
          type: string
          default: ICP-Brasil
        organizationUnits:
          type: array
          items:
            type: string
          description: Three organizationUnit RDNs (CA short name, CNPJ, validation method).
          example:
            - Example CA
            - "12345678000199"
            - Validacao por certificado digital
        responsiblePersonName:
          type: string
          description: ICP-Brasil otherName `2.16.76.1.3.2`.
        companyCnpj:
          type: string
          description: ICP-Brasil otherName `2.16.76.1.3.3`.
          example: "12345678000199"
        responsiblePersonData:
          type: string
          description: ICP-Brasil otherName `2.16.76.1.3.4` (21-char composite).
          example: "197001010000000000000"
        companyCei:
          type: string
          description: ICP-Brasil otherName `2.16.76.1.3.7`.
          example: "123456789012"
    QwacOpensslConfigRequest:
      type: object
      required:
        - commonName
        - country
        - organization
        - organizationIdentifier
        - dnsNames
      properties:
        commonName:
          type: string
          description: Subject commonName; typically the transport hostname.
          example: psp.example.com
        country:
          type: string
          minLength: 2
          maxLength: 2
          description: Two-letter ISO 3166-1 country code.
          example: GB
        stateOrProvince:
          type: string
        locality:
          type: string
        organization:
          type: string
          description: Legal-person registered name.
          example: Example PSP plc
        organizationIdentifier:
          type: string
          description: |
            ETSI EN 319 412-1 §5.1.4 organizationIdentifier. For PSD2 use the
            `PSD{country}-{NCA-Id}-{PSP-Id}` syntax from ETSI TS 119 495.
          example: PSDGB-FCA-123456
        serialNumber:
          type: string
          description: Optional subject serialNumber RDN.
        dnsNames:
          type: array
          items:
            type: string
          description: DNS subjectAltName entries; at least one is required.
          example:
            - psp.example.com
        psd2Roles:
          type: array
          items:
            type: string
            enum: [PSP_AS, PSP_PI, PSP_AI, PSP_IC]
          description: |
            Optional. When supplied the config emits the PSD2 qcStatement
            (`0.4.0.19495.2`) listing these PSP roles and the supervising
            NCA. Requires `ncaName` and `ncaId`.
        ncaName:
          type: string
          description: Supervising National Competent Authority name (PSD2).
          example: Financial Conduct Authority
        ncaId:
          type: string
          description: Supervising NCA identifier (PSD2).
          example: GB-FCA
        pdsLocations:
          type: array
          description: |
            Optional. PKI Disclosure Statement URLs, emitted as a `QcPDS`
            qcStatement.
          items:
            $ref: '#/components/schemas/PdsLocation'
    QsealOpensslConfigRequest:
      type: object
      required:
        - commonName
        - country
        - organization
        - organizationIdentifier
      properties:
        commonName:
          type: string
          example: PSP Seal
        country:
          type: string
          minLength: 2
          maxLength: 2
          example: GB
        stateOrProvince:
          type: string
        locality:
          type: string
        organization:
          type: string
          example: Example PSP plc
        organizationIdentifier:
          type: string
          example: PSDGB-FCA-123456
        serialNumber:
          type: string
        psd2Roles:
          type: array
          items:
            type: string
            enum: [PSP_AS, PSP_PI, PSP_AI, PSP_IC]
          description: |
            Optional. When supplied the config emits the PSD2 qcStatement.
            Requires `ncaName` and `ncaId`.
        ncaName:
          type: string
        ncaId:
          type: string
        pdsLocations:
          type: array
          items:
            $ref: '#/components/schemas/PdsLocation'
        onQscd:
          type: boolean
          default: false
          description: |
            When `true` the config emits the `QcSSCD` qcStatement, signalling
            that the private key is held on a Qualified Signature/Seal
            Creation Device.
...
    OpensslConfigResponse:
      type: object
      required: [filename, config]
      properties:
        filename:
          type: string
          description: |
            Suggested filename for saving the config; `brcac.cnf` for the
            BRCAC endpoint, `brseal.cnf` for the BRSEAL endpoint.
          example: brcac.cnf
        config:
          type: string
          format: byte
          description: |
            The generated openssl `req` config, base64-encoded. Decode and
            save to disk, then run
            `openssl req -new -config <file>.cnf -key <your-key>.pem -out csr.pem`
            to produce the CSR.
    IssuedCertificate:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Stable identifier of the persisted certificate.
        serialNumber:
          type: string
          description: The certificate serial number, hex-encoded.
          example: 1a2b3c4d
        subject:
          type: string
          description: The certificate subject distinguished name.
          example: CN=service.example.com
        certificate:
          type: string
          format: byte
          description: The signed leaf certificate, as base64-encoded PEM.
        chain:
          type: string
          format: byte
          description: The certificate chain, leaf-first, as base64-encoded PEM.
        csr:
          type: string
          format: byte
          description: |
            The PKCS#10 the certificate was issued from, as base64-encoded
            PEM — exactly what the caller submitted to `/v1/certificates`.
            Persisted for audit. Omitted only for certificates persisted
            before this field existed.
        notBefore:
          type: string
          format: date-time
          description: Start of the certificate validity window.
        notAfter:
          type: string
          format: date-time
          description: End of the certificate validity window.
        issuedAt:
          type: string
          format: date-time
          description: When the service issued the certificate.
...
  responses:
    BadRequest:
      description: |
        The request body or CSR was malformed. The `error` code is
        `invalid_request` for a bad body, or `bad_csr` for a CSR that could
        not be parsed as a PKCS#10 request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: The bearer token is missing, empty, or not active.
      headers:
        WWW-Authenticate:
          schema:
            type: string
          description: 'Bearer challenge, e.g. Bearer error="invalid_token", ...'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: The token is active but lacks the required scope.
      headers:
        WWW-Authenticate:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: No certificate exists with the given id.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerError:
      description: An unexpected error occurred while serving the request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'